requirements.txt
accelerate==0.16.0
albumentations~=1.3.0
bitsandbytes==0.35.4
diffusers==0.13.1
gitpython~=3.1.31
fastapi
ftfy~=6.1.1
modelcards~=0.1.6
tensorboard
tensorflow==2.11.0; sys_platform != 'darwin' or platform_machine != 'arm64'
tensorflow-macos==2.11.0; sys_platform == 'darwin' and platform_machine == 'arm64'
mediapipe-silicon; sys_platform == 'darwin'
mediapipe; sys_platform != 'darwin'
tqdm~=4.64.1
transformers~=4.26.1
discord-webhook~=1.1.0
lion-pytorch~=0.0.7
xformers==0.0.17.dev464
dadaptation==1.5.0